Move makevsix into gulpfile.js
