+++ package.json
@@ -24,6 +24,7 @@
     "semver": "4.3.3",
     "shelljs": "^0.3.0",
     "sync-request": "3.0.1",
+    "tfx-cli": "^0.3.40",
     "tslint": "^3.15.1",
     "typescript": "1.8.7",
     "validator": "3.33.0"
